chatbot_flask_requirements.txt
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
ChatterBot==1.0.0
chatterbot-corpus==1.2.0
click==7.1.2
contextvars==2.4
cymem==2.0.5
dataclasses==0.6
Flask==2.0.1
idna==2.10
immutables==0.15
importlib-metadata==4.5.0
importlib-resources==5.1.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
mathparse==0.1.2
murmurhash==1.0.5
nltk==3.6.2
numpy==1.19.5
packaging==20.9
pathy==0.5.2
Pint==0.17
preshed==3.0.5
pydantic==1.7.4
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2021.1
PyYAML==3.13
regex==2021.4.4
requests==2.25.1
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.6
SQLAlchemy==1.2.19
srsly==2.4.1
thinc==8.0.4
tqdm==4.61.1
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wasabi==0.8.2
Werkzeug==2.0.1
zipp==3.4.1
